[Refactor] Moved SvgIcon from the svg-icon folder to the root.
