+++ src/svg-icons/drop-down-arrow.jsx
@@ -1,5 +1,5 @@
 var React = require('react');
-var SvgIcon = require('./svg-icon');
+var SvgIcon = require('../svg-icon');
 
 var DropDownArrow = React.createClass({
