+++ src/svg-icons/toggle-radio-button-off.jsx
@@ -1,5 +1,5 @@
 var React = require('react');
-var SvgIcon = require('./svg-icon');
+var SvgIcon = require('../svg-icon');
 
 var RadioButtonOff = React.createClass({
